Modified addDrag.m and addCFD.m to reflect output format from CFD model
As Tyler understands it:
The array VelocitiesAtProbes contains vectored velocities at the center of each line segment and at the COM of any bodies that are not identified as turbines.
The array ForcesOnBodies contains force (column 1) and moment (column 2) magnitudes acting on bodies identified as turbines, but not other bodies such as buoys or whatevs.
Mooring model should now be compatible with this format.