modify the cosine correction thing for thrust forces
instead of applying the correction inside the Newton iterations
apply the correction before calling the Newton solver
and add loop around the UWMDNewton solver.  Now the mooring
model is called multiple times and the turbine forces are adjusted
iteratively. Added some figures to show the output, the mooring model
results look good, but something strange is happening with the CFD results